Merged KICK and PRIVMSG proofs of concept
They match correctly in the _processQueue function.
Need FD source to be accessible from the Message object
Also moved _checkPass() call to _processQueue()

I want to avoid a mass of if-else string comparisons. A switch based on
numeric equivalents of the commands should be faster and easier to read
I think.