Server can parse multiple messages in one chunk
Some clients send PASS NICK USER in a single lump and we were ignoring
the possibility of str_buf containing multiple \n delimited commands.
Now, we loop through str_buf deleting up to newlines until it is
cleared.

This is a dirty hack and needs to be tested on other clients - Hexchat
is happy now but Konversation never showed this problem anyway. Any regressions?