Channel: 2 getBroadcastFDs methods
One returns a list of  all FDs of Users in a channel, the other removes the sending FD from the list.

The idea is that this eases the creation of Messages that should go to a
channel (not a single user) by giving a method by which the targets can
be filled (correctly, we hope).
Note that we can expand this to exclude invisible or non-operators users
if needed.