ACommand: add pointer to the server
Without this, cannot access e.g. Channel or user listings from
commands(!)

I think that the pointer will be lightweight and not a memory danger. It
makes sense (to me) that a command is linked to the server it runs on,
but I don't see it as a "child" of that server so not inheriting.