Message parsing hardened against multi newlines
If a packet had more than one \n command, it was being parsed
incorrectly - most of the parameters were pulled into the first command,
some were lost.
To combat this:
- Message::parseMessage throws away everything after the first newline
- after calling makeMessage, the server loop throws away everything *before* the first newline
before checking if it should call again
(Note that there's a wee loop trying to handle all combinations of \r
and \n)
Also, disabled the exception thrown on short messages. And an exception
in the message set up no longer disconnects the client. (That was
excessive; the new state of affairs is not perfect though)