MESSAGES: consolidate Num reply parameters
Add one static method where are the reply numbers are case'd and return the
parameters to be used in the reply.
It uses defaults / nullables so we can ignore channel or User if we want.
The main benefit is having all of these in a single list and not spread
out across 3.