USER stores the Servername
This is mostly for compliance with GETWHOREPLY format.
We now include Server.hpp but instead of referring to it directly, we
store it per-user. It will always be the same! but this way we have a
path to acknowledging users (and channels, etc) that are from *other*
servers, if we wanted to extend the project that way.